+++ src/lib.rs
@@ -1,6 +1,9 @@
 //! TODO
+extern crate byteorder;
 extern crate crossbeam_channel;
+extern crate data_encoding;
 extern crate opentracingrust;
+extern crate rand;
 
 
 mod tracer;