+++ Changelog.md
@@ -1,6 +1,13 @@
+Roadmap
+=======
+- Support setting `SpanContext` sampled and debug state.
+
+
 Version 0.1.0 (Unreleased)
 ==========================
 
 Features
 --------
 - Initial library structure.
+- TraceID implementation.
+- Zipkin SpanContext implementation.